Repo successfully cloaned and debugged
I did the following to the original project:  1. I created the RAW_DATA_FOLDER, CLEAN_DATA_FOLDER, and TRANSFORM_DATA_FOLDER. 2. I changed the original  'URLs' of the folders to the ones I created. 3. I debugged the errors that popped up one after the other to get the code running.